♻️ Merge the sync provider into the new naming module *provider* to converge the relative logic.
