+++ test/contract_test/info_provider/mcp/test_ai_frameworks_contract.py
@@ -6,8 +6,7 @@ from typing import Any, Dict, List, Sequence
 import httpx
 import pytest
 
-from gearmeshing_ai.info_provider.mcp.provider import AsyncMcpClient
-from gearmeshing_ai.info_provider.mcp.client_sync import McpClient
+from gearmeshing_ai.info_provider.mcp.provider import AsyncMcpClient, McpClient
 from gearmeshing_ai.info_provider.mcp.schemas.config import (
     GatewayConfig,
     McpClientConfig,
