+++ test/unit_test/info_provider/mcp/test_policy_read_only.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, List
 import httpx
 import pytest
 
-from gearmeshing_ai.info_provider.mcp.client_sync import McpClient
+from gearmeshing_ai.info_provider.mcp.provider import McpClient
 from gearmeshing_ai.info_provider.mcp.errors import ToolAccessDeniedError
 from gearmeshing_ai.info_provider.mcp.policy import ToolPolicy
 from gearmeshing_ai.info_provider.mcp.schemas.config import (
